Rename package for auto-installer tool
